bitfield - 2 complete

+++ bitfield/src/lib.rs
@@ -11,16 +11,16 @@
 // From the perspective of a user of this crate, they get all the necessary APIs
 // (macro, trait, struct) through the one bitfield crate.
 pub use bitfield_impl::bitfield;
-use seq::seq;
+use seq_macro::seq;
 
 pub trait Specifier {
-  const BITS : u8;
+  const BITS : usize;
 }
 
 seq!(N in 0..64 {
   pub enum B#N {}
 
   impl Specifier for B#N{
-    const BITS : u8 = N;
+    const BITS : usize = N;
   }
 });
